HW 6 change update and delete comments

+++ src/repository/commentsRepository.ts
@@ -1,4 +1,4 @@
-import {CommentDBModalType} from "../types/types";
+import {CommentDBModalType, UserDBType} from "../types/types";
 import {commentsCollection} from "./db";
 
 
@@ -6,12 +6,12 @@ export const commentsRepository = {
     async createNewComments(newComment: CommentDBModalType) {
          return commentsCollection.insertOne(newComment)
     },
-    async updateComment(id: string, content: string){
-        const updateComment = await commentsCollection.updateOne({id}, {$set: {content}})
+    async updateComment(id: string, content: string, user:UserDBType){
+        const updateComment = await commentsCollection.updateOne({id, userId: user.id}, {$set: {content}})
         return updateComment.matchedCount === 1
     },
-    async deleteComment(id:string){
-        const deleteComment = await commentsCollection.deleteOne({id})
+    async deleteComment(id:string, user: UserDBType){
+        const deleteComment = await commentsCollection.deleteOne({id, userId: user.id})
         return deleteComment.deletedCount === 1
     }
 }
